+++ quantify_scheduler/backends/qblox/helpers.py
@@ -1,7 +1,6 @@
 # Repository: https://gitlab.com/quantify-os/quantify-scheduler
 # Licensed according to the LICENCE file on the main branch
 """Helper functions for Qblox backend."""
-
 import re
 from copy import deepcopy
 from collections import UserDict
@@ -40,7 +39,7 @@ def find_inner_dicts_containing_key(d: dict, key: Any) -> List[dict]:
     :
         A list containing all the inner dictionaries containing the specified key.
     """
-    dicts_found = list()
+    dicts_found = []
     if isinstance(d, dict):
         if key in d:
             dicts_found.append(d)
@@ -78,7 +77,7 @@ def find_all_port_clock_combinations(d: dict) -> List[Tuple[str, str]]:
         A list containing tuples representing the port and clock combinations found
         in the dictionary.
     """
-    port_clocks = list()
+    port_clocks = []
     dicts_with_port = find_inner_dicts_containing_key(d, "port")
     for inner_dict in dicts_with_port:
         if "port" in inner_dict:
@@ -429,7 +428,7 @@ def generate_port_clock_to_device_map(
         the same device.
     """
 
-    portclock_map = dict()
+    portclock_map = {}
     for device_name, device_info in hardware_cfg.items():
         if not isinstance(device_info, dict):
             continue
@@ -477,8 +476,8 @@ def assign_pulse_and_acq_info_to_devices(
 
     portclock_mapping = generate_port_clock_to_device_map(hardware_cfg)
 
-    for op_timing_constraint in schedule.schedulables.values():
-        op_hash = op_timing_constraint["operation_repr"]
+    for schedulable in schedule.schedulables.values():
+        op_hash = schedulable["operation_repr"]
         op_data = schedule.operations[op_hash]
 
         if isinstance(op_data, WindowOperation):
@@ -491,7 +490,7 @@ def assign_pulse_and_acq_info_to_devices(
                 f"Operation data: {repr(op_data)}"
             )
 
-        operation_start_time = op_timing_constraint["abs_time"]
+        operation_start_time = schedulable["abs_time"]
         for pulse_data in op_data.data["pulse_info"]:
             if "t0" in pulse_data:
                 pulse_start_time = operation_start_time + pulse_data["t0"]
@@ -500,8 +499,6 @@ def assign_pulse_and_acq_info_to_devices(
 
             port = pulse_data["port"]
             clock = pulse_data["clock"]
-            if port is None:
-                continue  # ignore idle pulses
 
             combined_data = OpInfo(
                 name=op_data.data["name"],
@@ -509,29 +506,41 @@ def assign_pulse_and_acq_info_to_devices(
                 timing=pulse_start_time,
             )
 
-            if (port, clock) not in portclock_mapping:
-                raise KeyError(
-                    f"Could not assign pulse data to device. The combination"
-                    f" of port {port} and clock {clock} could not be found "
-                    f"in hardware configuration.\n\nAre both the port and clock "
-                    f"specified in the hardware configuration?\n\nRelevant operation:\n"
-                    f"{combined_data}."
+            if port is None:
+                # Distribute clock operations to all sequencers utilizing that clock
+                for (map_port, map_clock), device_name in portclock_mapping.items():
+                    if map_clock == clock:
+                        device_compilers[device_name].add_pulse(
+                            port=map_port, clock=clock, pulse_info=combined_data
+                        )
+            else:
+                if (port, clock) not in portclock_mapping:
+                    raise KeyError(
+                        f"Could not assign pulse data to device. The combination "
+                        f"of port {port} and clock {clock} could not be found "
+                        f"in hardware configuration.\n\nAre both the port and clock "
+                        f"specified in the hardware configuration?\n\n"
+                        f"Relevant operation:\n{combined_data}."
+                    )
+                device_name = portclock_mapping[(port, clock)]
+                device_compilers[device_name].add_pulse(
+                    port=port, clock=clock, pulse_info=combined_data
                 )
-            dev = portclock_mapping[(port, clock)]
-            device_compilers[dev].add_pulse(port, clock, pulse_info=combined_data)
 
         for acq_data in op_data.data["acquisition_info"]:
             if "t0" in acq_data:
                 acq_start_time = operation_start_time + acq_data["t0"]
             else:
                 acq_start_time = operation_start_time
+
             port = acq_data["port"]
             clock = acq_data["clock"]
+
             if port is None:
                 continue
 
             hashed_dict = without(acq_data, ["t0", "waveforms"])
-            hashed_dict["waveforms"] = list()
+            hashed_dict["waveforms"] = []
             for acq in acq_data["waveforms"]:
                 hashed_dict["waveforms"].append(without(acq, ["t0"]))
 
@@ -540,27 +549,35 @@ def assign_pulse_and_acq_info_to_devices(
                 data=acq_data,
                 timing=acq_start_time,
             )
+
             if (port, clock) not in portclock_mapping:
                 raise KeyError(
-                    f"Could not assign acquisition data to device. The combination"
-                    f" of port {port} and clock {clock} could not be found "
+                    f"Could not assign acquisition data to device. The combination "
+                    f"of port {port} and clock {clock} could not be found "
                     f"in hardware configuration.\n\nAre both the port and clock "
-                    f"specified in the hardware configuration?\n\nRelevant operation:\n"
-                    f"{combined_data}."
+                    f"specified in the hardware configuration?\n\n"
+                    f"Relevant operation:\n{combined_data}."
                 )
-            dev = portclock_mapping[(port, clock)]
-            device_compilers[dev].add_acquisition(port, clock, acq_info=combined_data)
+            device_name = portclock_mapping[(port, clock)]
+            device_compilers[device_name].add_acquisition(
+                port=port, clock=clock, acq_info=combined_data
+            )
 
 
 def convert_hw_config_to_portclock_configs_spec(
     hw_config: Dict[str, Any],
-):
+) -> Dict[str, Any]:
     """
     Converts possibly old hardware configs to the new format introduced by
-    the new dynamic sequencer allocation feature. I.e. manual
-    assignment between sequencers and portclocks under each output is removed, and
-    instead only a list of port-clock configurations is specified,
-    under the new `portclock_configs` key.
+    the new dynamic sequencer allocation feature.
+
+    Manual assignment between sequencers and port-clock combinations under each output
+    is removed, and instead only a list of port-clock configurations is specified,
+    under the new ``"portclock_configs"`` key.
+
+    Furthermore, we scan for ``"latency_correction"`` defined at sequencer or
+    portclock_configs level and store under ``"port:clock"`` under toplevel
+    "latency_corrections" key.
 
     Parameters
     ----------
@@ -575,25 +592,33 @@ def convert_hw_config_to_portclock_configs_spec(
 
     """
 
+    def _update_hw_config(nested_dict, max_depth=4):
+        if max_depth == 0:
+            return
+        # List is needed because the dictionary keys are changed during recursion
+        for key, value in list(nested_dict.items()):
+            if isinstance(key, str) and re.match(r"^seq\d+$", key):
+                nested_dict["portclock_configs"] = nested_dict.get(
+                    "portclock_configs", []
+                )
+                # Move latency_corrections to parent level of hw_config
+                if "latency_correction" in value.keys():
+                    hw_config["latency_corrections"] = hw_config.get(
+                        "latency_corrections", {}
+                    )
+                    latency_correction_key = f"{value['port']}-{value['clock']}"
+                    hw_config["latency_corrections"][latency_correction_key] = value[
+                        "latency_correction"
+                    ]
+                    del value["latency_correction"]
+
+                nested_dict["portclock_configs"].append(value)
+                del nested_dict[key]
+
+            elif isinstance(value, dict):
+                _update_hw_config(value, max_depth - 1)
+
     hw_config = deepcopy(hw_config)
-
-    for device_info in hw_config.values():
-        if not isinstance(device_info, dict):
-            continue
-
-        for io, io_cfg in device_info.items():
-            if not isinstance(io_cfg, dict):
-                continue
-
-            new_io_cfg = {}
-            new_io_cfg["portclock_configs"] = io_cfg.get("portclock_configs", [])
-
-            for entry, value in io_cfg.items():
-                if not re.match("^seq\d+$", entry):
-                    new_io_cfg[entry] = value
-                    continue
-                new_io_cfg["portclock_configs"].append(io_cfg[entry])
-
-            device_info[io] = new_io_cfg
+    _update_hw_config(hw_config)
 
     return hw_config