+++ CHANGELOG.rst
@@ -7,9 +7,16 @@ Unreleased
 * Compilation - Added `acq_protocol` optional parameter to the `Measure` gate. (!386)
 * QuantumDevice - Added implementation for `edges` in the quantum device config in order to support two qubit operations. The `Edge` has been added as an abstract base class for edges to be added to a device. (!389)
 * Qblox backend - Fix for supplying negative NCO phase (!393)
+* Qblox backend - Fix compilation of ShiftClockPhase (!404, broken by merge of !328)
 * Qblox backend - Fix for outputting signals on even output paths of qblox hardware in real_output_x mode (!397)
-* Operation - Added a `symmetric` key in the `gate_info` to flag symmetric operations. (!389)
+* Qblox backend - Fix cluster compatibility when converting old hwconfig to new specs (!419)
+* Qblox backend - Latency corrections must now be defined in top layer of hw config (!400)
+* Qblox backend - Fix combination of cluster and latency corrections when converting hw_configs to new specs  (!417)
+* Qblox backend - Fix handling of composite pulses (#299, !411)
+* Operations - Added a `symmetric` key in the `gate_info` to flag symmetric operations. (!389)
+* Operations - Introduce basic CZ-gate via `CompositeSquareEdge` (utilizing `quantify_scheduler.operations.pulse_factories.composite_square_pulse`); replaces the incomplete `SuddenNetZeroEdge` basic CZ-gate implementation (!411)
 * Helpers - Moved MockLocalOscillator definition from tests to `helpers.mock_instruments.MockLocalOscillator` (!392, !336).
+* DRAG Pulse - Removed an extra G_amp factor from the Q component (derivative pulse). (#298, !406)
 
 Breaking changes
 ~~~~~~~~~~~~~~~~